% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_1.R
\name{get_daily_FIPS_averages}
\alias{get_daily_FIPS_averages}
\title{This function retrieves daily averages for the specified county codes or the
state. Start and end dates are optional inputs.}
\usage{
get_daily_FIPS_averages(
  county_FIPS_codes,
  measurements,
  path = NULL,
  coverage = 1,
  plot_directory = NULL,
  start_date = NULL,
  end_date = NULL,
  verbose = NULL
)
}
\arguments{
\item{county_FIPS_codes}{An integer array of county FIPS codes.}

\item{measurements}{Four measurements currently available:
Temperature (Average, Minimum, Maximum) and Precipitation. Defaults to
c("Temp_avg","Temp_min", "Temp_max", "Precipitation").}

\item{path}{Relative path to the folder where the retrieved data will be
stored. Defaults to a folder named EpiNOAA_call. TODO}

\item{coverage}{Shows the fraction of data that is not missing. Defaults to 1.}

\item{plot_directory}{Relative path to the folder where the time series plot
of the daily data will be saved. Defaults to a folder named EpiNOAA_plots. TODO}
}
\value{
Basic summary statistics.
@examples
get_daily_averages(37001,c("Temp_avg","Temp_min", "Temp_max", "Precipitation"))
get_daily_averages(37001,c("Temp_avg","Temp_min", "Temp_max", "Precipitation"), start_date = "2018-09-21", end_date = "2019-09-21")
}
\description{
This function retrieves daily averages for the specified county codes or the
state. Start and end dates are optional inputs.
}
