% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_1.R
\name{get_historic_data}
\alias{get_historic_data}
\title{This function retrieves historic daily averages for a given date.}
\usage{
get_historic_data(daily_temperatures, historic_date)
}
\arguments{
\item{daily_temperatures}{A dataframe from the function get_daily_FIPS_averages().}

\item{historic_date}{A date for which the}
}
\value{
A dataframe containing
@examples
get_historic_data(daily_temperatures, "2019-09-21")
}
\description{
This function retrieves historic daily averages for a given date.
}
