% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_1.R
\name{get_daily_FIPS_averages}
\alias{get_daily_FIPS_averages}
\title{This function retrieves daily averages for the specified county codes or the
state. Start and end dates are optional inputs.}
\usage{
get_daily_FIPS_averages(
  county_FIPS_codes,
  measurements,
  dailyTemperatures,
  coverage = 1,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{county_FIPS_codes}{An integer array of county FIPS codes.}

\item{measurements}{Four measurements currently available:
Temperature (Average, Minimum, Maximum) and Precipitation. Defaults to
c("Temp_avg","Temp_min", "Temp_max", "Precipitation").}

\item{coverage}{Shows the fraction of data that is not missing. Defaults to 1.}

\item{start_date}{Indicates the start date}

\item{end_date}{Indicates the end date}
}
\value{
Basic summary statistics.
@examples
get_daily_averages(37001,c("Temp_avg","Temp_min", "Temp_max", "Precipitation"))
get_daily_averages(37001,c("Temp_avg","Temp_min", "Temp_max", "Precipitation"), start_date = "2018-09-21", end_date = "2019-09-21")
}
\description{
This function retrieves daily averages for the specified county codes or the
state. Start and end dates are optional inputs.
}
