% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_1.R
\name{get_data_dictionary}
\alias{get_data_dictionary}
\title{This function retrieves daily averages for the specified county codes or the
state. Start and end dates are optional inputs.}
\usage{
get_data_dictionary()
}
\arguments{
\item{input}{No input required}
}
\value{
A list with two elements county and state level with available
meta-data, start and end dates and coverage for each measurement.
@examples
get_daily_averages()
}
\description{
This function retrieves daily averages for the specified county codes or the
state. Start and end dates are optional inputs.
}
